Add input to run, sepearate commit for git conflict
The input argument is passed to Popen.communicate() and thus to the subprocess’s stdin. If used it must be a byte sequence, or a string if encoding or errors is specified or text is true. When used, the internal Popen object is automatically created with stdin set to PIPE, and the stdin argument may not be used as well.

Co-authored-by: Marcos Diez <marcos@unitron.com.br>

Co-authored-by: Shasheen Ediriweera <1414379+shasheene@users.noreply.github.com>