Fix GPG package for end-of-life Bionic and Focal
Downloads the GPG public key needed for package verification from the
fingerprints, in order to fix the GPG package verification of Ubuntu 18.04
(Bionic) and Ubuntu 20.04 (Focal).

The failed verification issue was already observed in Ubuntu 18.04 (Bionic) in
December 2024, so signature check was temporarily disabled (#538). The same
issue has now occurred on the Focal build.

Further investigation revealed the manual import for package verification was
all that was needed to resolve this.

Issue: #538