it: Remove unnecessary 'reset' of VM disk in tests
Removes the unneeded integration-test reset call which calls a function that
copies over a blank HD, the subsequent deploy does an rsync command to achieve
the same thing so the reset does nothing effectual right now.

And it should speed things up to not copy over the blank disk first.